project_name/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-taggit==3.0.0
Markdown==3.4.1
Pillow==9.2.0
sqlparse==0.4.3
tzdata==2022.5
django-allauth==0.51.0